Completed re factoring url handling into an object.
Plenty of work done in this commit. Everywhere that used to accept a url
string now excepts the \Proem\IO\Url object instead. This object is then
reponsible for parsing a url into the various tokens / parts.

There is still a bit of work to be done within the url class itself in
regards to validation. I would also like the url object to be able to
(probably via proxying to the router) reverse route a given
controller/action into a url string. The validation part is important,
the reverse routing is candy for now.

Lots of classes & tests where touched in this commit as a url is quite
obviously a central part of the dispatch process.