Implement Tidy up Chain
This implements a number of new features and removes the AbstractChain
completely from Proem. Features include:

 * The ability to inject events into a specific location within the Chain
   of Events. injectEvents()

 * The ability to seek the previous event within the Chain. getPreviousEvent()

 * Fixed the registerEvents() method to no longer use a loop.

The testCanRegisterSingleEvent test has been commented out within the
ChainTest test case as it is now throwing an error due to the abstract
mock and type hinting. This issue didn't exist prior to these changes,
however I am unable to resolv the issue at this point in time, while the
tests will need a great deal of refactoring in the near future anyway
due to issues #28 & #30.

This work fixes #25